FitHub/tools/affiliate_link_generator/requirements.txt
firebase-admin>=6.3.0
python-dotenv>=1.0.1
